+++ packages/author-profile/src/head.web.js
@@ -5,11 +5,15 @@ import { renderTreeArrayAsText } from "@times-components/markup-forest";
 
 import { propTypes as authorProfileHeadPropTypes } from "./author-profile-head-prop-types";
 
-function Head({ description, name }) {
-  const content =
-    description && description.length
-      ? renderTreeArrayAsText(description).substring(0, 200)
-      : `Get up to date information and read all the latest articles from ${name}.`;
+function Head({ metaDescription, description, name }) {
+  let content = `Get up to date information and read all the latest articles from ${name}.`;
+
+  if (metaDescription) {
+    content = metaDescription;
+  } else if (description && description.length) {
+    content = renderTreeArrayAsText(description).substring(0, 200);
+  }
+
   return (
     <Helmet>
       <title>{name} | The Times &amp; The Sunday Times</title>
@@ -19,8 +23,13 @@ function Head({ description, name }) {
 }
 
 Head.propTypes = {
+  metaDescription: PropTypes.string,
   description: authorProfileHeadPropTypes.biography.isRequired,
   name: PropTypes.string.isRequired
+};
+
+Head.defaultProps = {
+  metaDescription: null
 };
 
 export default Head;