+++ packages/article-paragraph/CHANGELOG.md
@@ -3,6 +3,54 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.4.217](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.216...@times-components/article-paragraph@1.4.217) (2021-08-27)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.216](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.215...@times-components/article-paragraph@1.4.216) (2021-08-27)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.215](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.214...@times-components/article-paragraph@1.4.215) (2021-08-27)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.214](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.213...@times-components/article-paragraph@1.4.214) (2021-08-27)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.213](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.212...@times-components/article-paragraph@1.4.213) (2021-08-27)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
+## [1.4.212](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.211...@times-components/article-paragraph@1.4.212) (2021-08-25)
+
+**Note:** Version bump only for package @times-components/article-paragraph
+
+
+
+
+
 ## [1.4.211](https://github.com/newsuk/times-components/compare/@times-components/article-paragraph@1.4.210...@times-components/article-paragraph@1.4.211) (2021-08-24)
 
 **Note:** Version bump only for package @times-components/article-paragraph